# sysquare-project the project work for sysquare i have create a registration page using html css javascript and bootstrap.
the data that is inserted is stored locally in the browser localStorage

in this project the user inputs the given data and the data is stored locally after submittng the form.

the data can be updated/deleted with the help of the button given below the  submit button.

to update the data the user has to given the key and the value of the key he/she want to update. to give the input i have used prompt() amd stored the value from prompt to the local variable.

to delete the stored data the user has to just give the key of the data stored and the key is asked with the help of prompt and the data is passed to the local variable to pass on to the next if else condition.